🐛 fix(chat): gate single-step delete on heterogeneous-agent tail error
Address review: the special case applied to ANY errored child in any
assistantGroup, so a normal grouped reply that ended in a generic
tool/provider error would also drop down to single-block delete and leave
the rest of the turn behind. Narrow it to the actual scenario — the LAST
child being a heterogeneous-agent (CC/Codex) status error — using the
existing isHeterogeneousAgentStatusGuideError guard.

Extract that guard (+ its error-code set) into a dependency-free
Error/heterogeneous module so the action bar can import it without pulling
the whole Error UI barrel; index.tsx re-exports it to keep existing
consumers working.
